[docs] Split test ad networks
